context: model states & auto-(re)connect
this allows UIs to visualize the state (e.g. in a statusbar) or more
specifically handle scenarios where the context breaks and the user
needs informing and the ability to retry the connection

to facilitate these use cases a new state property is available that
fairly closely matches the internal state_t with the added functionality
that the Context state is also Connecting before we have an `m_context`
internally.

additionally a new autoConnecting property is available to track the
process of internal auto connect behavior on context loss. this is an
overarching state in which all other states may appear but may be
ignored in the UI on account of possibly being transient