context: always synthesize an event streamrestore
doesn't look like this is going to happen upstream in pipewire so let's
do it on a library level for convenience reasons.

when starting a new user profile there will initially be no stream
restores and by extension no event stream control in plasma-pa. to
prevent this from happening we now initially create a dummy restore that
will eventually be updated into the correct object once the real info
arrives.

this now also requires that we explicitly reset the restores on errors
as otherwise we'd have our synthesized restore but stream restoring in
general doesn't work.

BUG: 425974