Expose support for "Pro Audio" profile to clients
This audio profile is documented [1] to only make sense for truly "pro"
audio cards with 9+ channels. Expose a property for this so that clients
can filter it out of their profile lists in cases where that makes
sense.

Automatically filtering it out in Profile::availability() was
considered as an alternative method, but rejected because it would deny
clients the fine-grained control they may need (e.g. to not filter it
out for inactive cards, where it may be the only "turn it on" profile).

[1] https://gitlab.freedesktop.org/pipewire/pipewire/-/wikis/FAQ#what-is-the-pro-audio-profile